Incident-22111 Remove token from logs (#8)
## Context

Incident: https://app.datadoghq.com/incidents/22111

We don't want to log the tokens even when debugging. These debug logs
end up showing up in GitLab and k8s, and they contain real usable
tokens. We don't want to leak these tokens accidentally.

We remove the logging of the tokens, since that appears to be where the
leak is.