Incident-22111 Use initial successful response (#9)
## Context

Incident: https://app.datadoghq.com/incidents/22111

We noticed that in some cases, the initial call to the registry seems to
work, but we were ignoring the response we received instead of using it.
We change it so we now use the response we received.

However just because we received a response, doesn't mean that it has
useful data in it. On the other end, we have to check if the response
actually has a token in it. If it doesn't, we don't do anything with it.

## Testing

We should add a test to validate that using a registry that succeeds on
the initial request works.